# Gestion-Rh-UI/src/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
    "title": "OpenApi specification - Bassim",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
    "name": "bassim",
    "url": "https://bassimtabbe.com/",
    "email": "Bassim.tabbeb@gmail.com"
    },
    "license": {
    "name": "Licence name",
    "url": "https://some-url.com"
    },
    "version": "1.0"
    },
    "servers": [
    {
    "url": "http://localhost:8081/api/v1",
    "description": "Local ENV"
    },
    {
    "url": "https://bassimtabbe.com/e",
    "description": "PROD ENV"
    }
    ],
    "security": [
    {
    "bearerAuth": []
    }
    ],
    "paths": {
    "/utilisateur/{id}": {
    "get": {
    "tags": [
    "Utilisateur"
    ],
    "operationId": "findById",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UtilisateurDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "Utilisateur"
    ],
    "operationId": "update",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UtilisateurDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UtilisateurDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "Utilisateur"
    ],
    "operationId": "delete",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/formation/{id}": {
    "get": {
    "tags": [
    "Formation"
    ],
    "operationId": "findById_1",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/FormationDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "Formation"
    ],
    "operationId": "update_1",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/FormationDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/FormationDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "Formation"
    ],
    "operationId": "delete_1",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/event/{id}": {
    "get": {
    "tags": [
    "Event"
    ],
    "operationId": "findById_2",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/EventDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "Event"
    ],
    "operationId": "update_2",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/EventDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/EventDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "Event"
    ],
    "operationId": "delete_2",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/department/{id}": {
    "get": {
    "tags": [
    "Departement"
    ],
    "operationId": "findById_3",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DepartementDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "Departement"
    ],
    "operationId": "update_3",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DepartementDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DepartementDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "Departement"
    ],
    "operationId": "delete_3",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/demande/{id}": {
    "get": {
    "tags": [
    "Demande"
    ],
    "operationId": "findById_4",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DemandeDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "Demande"
    ],
    "operationId": "update_4",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DemandeDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DemandeDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "Demande"
    ],
    "operationId": "delete_4",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/acompte/{id}": {
    "get": {
    "tags": [
    "Acompte"
    ],
    "operationId": "findById_5",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AcompteDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "Acompte"
    ],
    "operationId": "update_5",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AcompteDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AcompteDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "Acompte"
    ],
    "operationId": "delete_5",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/TravailSup/{id}": {
    "get": {
    "tags": [
    "AutorisationTravailSup"
    ],
    "operationId": "findById_6",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AutorisationTravailSupDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "AutorisationTravailSup"
    ],
    "operationId": "update_6",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AutorisationTravailSupDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AutorisationTravailSupDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "AutorisationTravailSup"
    ],
    "operationId": "delete_6",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/Teletravail/{id}": {
    "get": {
    "tags": [
    "AutorisationTeletravail"
    ],
    "operationId": "findById_7",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AutorisationTeletravailDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "AutorisationTeletravail"
    ],
    "operationId": "update_7",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AutorisationTeletravailDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AutorisationTeletravailDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "AutorisationTeletravail"
    ],
    "operationId": "delete_7",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/Pret/{id}": {
    "get": {
    "tags": [
    "Pret"
    ],
    "operationId": "findById_8",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PretDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "Pret"
    ],
    "operationId": "update_8",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PretDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PretDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "Pret"
    ],
    "operationId": "delete_8",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/OrdreMission/{id}": {
    "get": {
    "tags": [
    "OrdreMission"
    ],
    "operationId": "findById_9",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/OrdreMissionDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "OrdreMission"
    ],
    "operationId": "update_9",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/OrdreMissionDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/OrdreMissionDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "OrdreMission"
    ],
    "operationId": "delete_9",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/Conge/{id}": {
    "get": {
    "tags": [
    "Conge"
    ],
    "operationId": "findById_10",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CongeDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "Conge"
    ],
    "operationId": "update_10",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CongeDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/CongeDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "Conge"
    ],
    "operationId": "delete_10",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/ChangementH/{id}": {
    "get": {
    "tags": [
    "ChangementHoraire"
    ],
    "operationId": "findById_11",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ChangementHoraireDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "ChangementHoraire"
    ],
    "operationId": "update_11",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ChangementHoraireDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ChangementHoraireDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "ChangementHoraire"
    ],
    "operationId": "delete_11",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
    "/AutorisationS/{id}": {
    "get": {
    "tags": [
    "AutorisationSortie"
    ],
    "operationId": "findById_12",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully retrieve entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AutorisationSortieDto"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "AutorisationSortie"
    ],
    "operationId": "update_12",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AutorisationSortieDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successfully updated entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/AutorisationSortieDto"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "AutorisationSortie"
    ],
    "operationId": "delete_12",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successfully delete entity",
    "content": {
    "application/json": {
    "schema": {
    "type": "string"
    }
    }
    }
    }
    }
    }
    },
...
    "/event": {
    "post": {
    "tags": [
    "Event"
    ],
    "operationId": "add_2",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/EventDto"
    }
    }
    },
    "required": true
    },
    "responses": {
    "201": {
    "description": "Successfully created entity",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/EventDto"
    }
    }
    }
    }
    }
    }
    },
    "/event/addEmploye/{id}": {
    "post": {
    "tags": [
    "Event"
    ],
    "operationId": "addEmployeToEvent",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "type": "integer",
    "format": "int64"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "OK"
    }
    }
    }
    },
...
    },
    "components": {
    "schemas": {
...
    "DepartementDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "nom": {
    "type": "string"
    },
    "manager": {
    "$ref": "#/components/schemas/Utilisateur"
    },
    "listEmploye": {
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/Utilisateur"
    }
    }
    }
    },
    "Role": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "name": {
    "type": "string"
    }
    }
    },
...
    "UtilisateurDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "nom": {
    "type": "string"
    },
    "prenom": {
    "type": "string"
    },
    "departement": {
    "$ref": "#/components/schemas/DepartementDto"
    },
    "cin": {
    "type": "string"
    },
    "service": {
    "type": "string"
    },
    "direction": {
    "type": "string"
    },
    "sexe": {
    "type": "string"
    },
    "adresse": {
    "type": "string"
    },
    "telephone": {
    "type": "string"
    },
    "img": {
    "type": "string"
    },
    "email": {
    "type": "string"
    },
    "username": {
    "type": "string"
    },
    "password": {
    "type": "string"
    },
    "accountLocked": {
    "type": "boolean"
    },
    "enabled": {
    "type": "boolean"
    },
    "roles": {
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/Role"
    }
    },
    "dembauche": {
    "type": "string",
    "format": "date-time"
    },
    "ejuridic": {
    "type": "string"
    }
    }
    },
    "FormationDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "dateD": {
    "type": "string",
    "format": "date-time"
    },
    "titre": {
    "type": "string"
    },
    "dateF": {
    "type": "string",
    "format": "date-time"
    },
    "duree": {
    "type": "integer",
    "format": "int32"
    },
    "lieu": {
    "type": "string"
    },
    "nbrPlace": {
    "type": "integer",
    "format": "int32"
    },
    "img": {
    "type": "string"
    },
    "description": {
    "type": "string"
    },
    "listEmploye": {
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/Utilisateur"
    }
    }
    }
    },
    "EventDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "dateD": {
    "type": "string",
    "format": "date-time"
    },
    "dateF": {
    "type": "string",
    "format": "date-time"
    },
    "duree": {
    "type": "integer",
    "format": "int32"
    },
    "lieu": {
    "type": "string"
    },
    "nbrPlace": {
    "type": "integer",
    "format": "int32"
    },
    "img": {
    "type": "string"
    },
    "titre": {
    "type": "string"
    },
    "description": {
    "type": "string"
    },
    "listEmploye": {
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/Utilisateur"
    }
    }
    }
    },
    "DemandeDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "statut": {
    "type": "string",
    "enum": [
    "Validee",
    "En_attente",
    "Refusée",
    "Annuler"
    ]
    },
    "type": {
    "type": "string"
    },
    "utilisateur": {
    "$ref": "#/components/schemas/Utilisateur"
    },
    "departement": {
    "$ref": "#/components/schemas/Departement"
    }
    }
    },
    "AcompteDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "statut": {
    "type": "string",
    "enum": [
    "Validee",
    "En_attente",
    "Refusée",
    "Annuler"
    ]
    },
    "type": {
    "type": "string"
    },
    "utilisateur": {
    "$ref": "#/components/schemas/Utilisateur"
    },
    "departement": {
    "$ref": "#/components/schemas/Departement"
    },
    "montant": {
    "type": "integer",
    "format": "int32"
    }
    }
    },
    "AutorisationTravailSupDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "statut": {
    "type": "string",
    "enum": [
    "Validee",
    "En_attente",
    "Refusée",
    "Annuler"
    ]
    },
    "type": {
    "type": "string"
    },
    "utilisateur": {
    "$ref": "#/components/schemas/Utilisateur"
    },
    "departement": {
    "$ref": "#/components/schemas/Departement"
    },
    "date": {
    "type": "string",
    "format": "date-time"
    },
    "motif": {
    "type": "string"
    },
    "hacces": {
    "type": "string",
    "format": "date-time"
    },
    "hsortie": {
    "type": "string",
    "format": "date-time"
    }
    }
    },
    "AutorisationTeletravailDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "statut": {
    "type": "string",
    "enum": [
    "Validee",
    "En_attente",
    "Refusée",
    "Annuler"
    ]
    },
    "type": {
    "type": "string"
    },
    "utilisateur": {
    "$ref": "#/components/schemas/Utilisateur"
    },
    "departement": {
    "$ref": "#/components/schemas/Departement"
    },
    "debut": {
    "type": "string",
    "format": "date-time"
    },
    "fin": {
    "type": "string",
    "format": "date-time"
    },
    "contact": {
    "type": "string"
    },
    "telephone": {
    "type": "string"
    },
    "lieu": {
    "type": "string"
    },
    "tache": {
    "type": "string"
    }
    }
    },
    "PretDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "statut": {
    "type": "string",
    "enum": [
    "Validee",
    "En_attente",
    "Refusée",
    "Annuler"
    ]
    },
    "type": {
    "type": "string"
    },
    "utilisateur": {
    "$ref": "#/components/schemas/Utilisateur"
    },
    "departement": {
    "$ref": "#/components/schemas/Departement"
    },
    "montant": {
    "type": "integer",
    "format": "int32"
    },
    "motif": {
    "type": "string"
    },
    "credit": {
    "type": "string"
    },
    "remboursement": {
    "type": "string"
    }
    }
    },
    "OrdreMissionDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "statut": {
    "type": "string",
    "enum": [
    "Validee",
    "En_attente",
    "Refusée",
    "Annuler"
    ]
    },
    "type": {
    "type": "string"
    },
    "utilisateur": {
    "$ref": "#/components/schemas/Utilisateur"
    },
    "departement": {
    "$ref": "#/components/schemas/Departement"
    },
    "dateS": {
    "type": "string",
    "format": "date-time"
    },
    "dateR": {
    "type": "string",
    "format": "date-time"
    },
    "motif": {
    "type": "string"
    },
    "hsortie": {
    "type": "string",
    "format": "date-time"
    },
    "hretour": {
    "type": "string",
    "format": "date-time"
    }
    }
    },
    "CongeDto": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int64"
    },
    "createdBy": {
    "type": "string"
    },
    "lastModifiedBy": {
    "type": "string"
    },
    "createdDate": {
    "type": "string",
    "format": "date-time"
    },
    "lastModifiedDate": {
    "type": "string",
    "format": "date-time"
    },
    "statut": {
    "type": "string",
    "enum": [
    "Validee",
    "En_attente",
    "Refusée",
    "Annuler"
    ]
    },
    "type": {
    "type": "string"
    },
    "utilisateur": {
    "$ref": "#/components/schemas/Utilisateur"
    },
    "departement": {
    "$ref": "#/components/schemas/Departement"
    },
    "duree": {
    "type": "integer",
    "format": "int32"
    },
    "debut": {
    "type": "string",
    "format": "date-time"
    },
    "fin": {
    "type": "string",
    "format": "date-time"
    },
    "reprise": {
    "type": "string",
    "format": "date-time"
    }
    }
    },
...
    },
    "securitySchemes": {
    "bearerAuth": {
    "type": "http",
    "description": "JWT auth description",
    "in": "header",
    "scheme": "bearer",
    "bearerFormat": "JWT"
    }
    }
    }
    }